Minor refactor. Move Postgres specific methods/logic into a PostgresUtils class, to avoid cluttering DatabaseUtils (as much)
